requirements.txt
# ⚡ Energy Agent - Requirements.txt
# ======================================
# Dépendances Python pour l'application Energy Agent
# Architecture LangGraph + Streamlit + DuckDB + Prophet

# 🐍 Python Core
# Python 3.11+ requis

# 📊 Data Science & Analytics
pandas>=2.0.0
numpy>=1.24.0
polars>=0.20.0

# 🗄️ Database
duckdb>=0.9.0

# 🎨 Visualization
plotly>=5.15.0
matplotlib>=3.7.0

# 🌐 Web Framework
streamlit>=1.28.0

# 🤖 AI & Machine Learning
google-generativeai>=0.3.0
langgraph>=0.1.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.1.0
langchain-community>=0.1.0
langchain-experimental>=0.1.0

# 🔮 Forecasting
prophet>=1.1.0

# 🔧 Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# 📅 Date & Time
python-dateutil>=2.8.0

# 🔍 Type Checking (optional)
typing-extensions>=4.5.0

# 📝 Logging
loguru>=0.7.0

# 🧪 Testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0

# 🔒 Security
cryptography>=41.0.0

# 📦 Packaging
setuptools>=68.0.0
wheel>=0.41.0

# 🚀 Performance
numba>=0.57.0
cython>=3.0.0
